fix 5->7 db migration
